@@ -3,67 +3,55 @@
   "license": "GPL-3.0-or-later",
   "version": "0.1.0",
   "dependencies": {
-    "@babel/core": "7.14.6",
-    "@monsonjeremy/react-leaflet": "^3.2.2",
-    "@pmmmwh/react-refresh-webpack-plugin": "0.4.3",
+    "@babel/core": "7.16.7",
+    "@babel/runtime": "^7.16.7",
+    "@pmmmwh/react-refresh-webpack-plugin": "0.5.4",
+    "@svgr/webpack": "^6.1.2",
     "@types/react-toastify": "^4.1.0",
     "@typescript-eslint/eslint-plugin": "^5.8.1",
     "@typescript-eslint/parser": "^5.8.1",
-    "babel-eslint": "^10.1.0",
     "babel-loader": "8.2.3",
-    "babel-plugin-named-asset-import": "^0.3.7",
     "babel-preset-react-app": "^10.0.1",
     "bfj": "^7.0.2",
     "case-sensitive-paths-webpack-plugin": "2.4.0",
-    "css-loader": "5.2.6",
+    "css-loader": "6.5.1",
+    "css-minimizer-webpack-plugin": "^3.3.1",
     "debug": "^4.3.3",
     "dotenv": "10.0.0",
     "dotenv-expand": "5.1.0",
-    "eslint": "^7.29.0",
-    "eslint-config-airbnb-typescript": "^12.3.1",
+    "eslint": "^8.6.0",
+    "eslint-config-airbnb-typescript": "^16.1.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.7.2",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-testing-library": "^4.6.0",
-    "eslint-webpack-plugin": "^2.5.4",
+    "eslint-config-react-app": "^7.0.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-webpack-plugin": "^3.1.1",
     "file-loader": "6.2.0",
     "fs-extra": "^10.0.0",
-    "html-webpack-plugin": "4.5.0",
+    "html-webpack-plugin": "5.5.0",
     "leaflet": "^1.7.1",
-    "mini-css-extract-plugin": "0.11.3",
-    "optimize-css-assets-webpack-plugin": "5.0.4",
-    "pnp-webpack-plugin": "1.6.4",
-    "postcss-flexbugs-fixes": "4.2.1",
-    "postcss-loader": "3.0.0",
-    "postcss-normalize": "8.0.1",
-    "postcss-preset-env": "6.7.0",
-    "postcss-safe-parser": "5.0.2",
-    "prettier": "^2.3.2",
+    "mini-css-extract-plugin": "2.4.5",
+    "prettier": "^2.5.1",
     "react": "^17.0.2",
-    "react-dev-utils": "^11.0.3",
+    "react-dev-utils": "^12.0.0",
     "react-dom": "^17.0.2",
-    "react-refresh": "^0.10.0",
-    "react-router-dom": "^5.2.0",
+    "react-leaflet": "^3.2.2",
+    "react-refresh": "^0.11.0",
+    "react-router-dom": "^6.2.1",
     "react-toastify": "^8.1.0",
     "resolve": "1.20.0",
-    "resolve-url-loader": "^3.1.2",
-    "semantic-ui-css": "^2.4.1",
+    "resolve-url-loader": "^4.0.0",
     "semantic-ui-react": "^2.0.4",
     "semver": "7.3.5",
-    "style-loader": "2.0.0",
+    "source-map-loader": "^3.0.0",
+    "style-loader": "3.3.1",
     "terrene-types": "file:../types",
-    "terser-webpack-plugin": "4.2.3",
-    "ts-pnp": "1.2.0",
-    "url-loader": "4.1.1",
-    "webpack": "4.44.2",
-    "webpack-dev-server": "3.11.2",
-    "webpack-manifest-plugin": "2.2.0",
-    "workbox-webpack-plugin": "5.1.4"
+    "terser-webpack-plugin": "5.3.0",
+    "webpack": "5.65.0",
+    "webpack-dev-server": "4.7.2",
+    "webpack-manifest-plugin": "4.0.2",
+    "workbox-webpack-plugin": "6.4.2"
   },
   "scripts": {
     "start": "node scripts/start.js",
