@@ -8,7 +8,7 @@ process.env.NODE_ENV = 'production';
 // ignoring them. In the future, promise rejections that are not handled will
 // terminate the Node.js process with a non-zero exit code.
 process.on('unhandledRejection', err => {
-  throw err;
+    throw err;
 });
 
 // Ensure environment variables are read.
@@ -40,7 +40,7 @@ const isInteractive = process.stdout.isTTY;
 
 // Warn and crash if required files are missing
 if (!checkRequiredFiles([paths.appHtml, paths.appIndexJs])) {
-  process.exit(1);
+    process.exit(1);
 }
 
 const argv = process.argv.slice(2);
@@ -54,162 +54,164 @@ const config = configFactory('production');
 const { checkBrowsers } = require('react-dev-utils/browsersHelper');
 checkBrowsers(paths.appPath, isInteractive)
     .then(() => {
-      // First, read the current file sizes in build directory.
-      // This lets us display how much they changed later.
-      return measureFileSizesBeforeBuild(paths.appBuild);
+        // First, read the current file sizes in build directory.
+        // This lets us display how much they changed later.
+        return measureFileSizesBeforeBuild(paths.appBuild);
     })
     .then(previousFileSizes => {
-      // Remove all content but keep the directory so that
-      // if you're in it, you don't end up in Trash
-      fs.emptyDirSync(paths.appBuild);
-      // Merge with the public folder
-      copyPublicFolder();
-      // Start the webpack build
-      return build(previousFileSizes);
+        // Remove all content but keep the directory so that
+        // if you're in it, you don't end up in Trash
+        fs.emptyDirSync(paths.appBuild);
+        // Merge with the public folder
+        copyPublicFolder();
+        // Start the webpack build
+        return build(previousFileSizes);
     })
     .then(
         ({ stats, previousFileSizes, warnings }) => {
-          if (warnings.length) {
-            console.log(chalk.yellow('Compiled with warnings.\n'));
-            console.log(warnings.join('\n\n'));
-            console.log(
-                '\nSearch for the ' +
-                chalk.underline(chalk.yellow('keywords')) +
-                ' to learn more about each warning.'
-            );
-            console.log(
-                'To ignore, add ' +
-                chalk.cyan('// eslint-disable-next-line') +
-                ' to the line before.\n'
-            );
-          } else {
-            console.log(chalk.green('Compiled successfully.\n'));
-          }
+            if (warnings.length) {
+                console.log(chalk.yellow('Compiled with warnings.\n'));
+                console.log(warnings.join('\n\n'));
+                console.log(
+                    '\nSearch for the ' +
+                    chalk.underline(chalk.yellow('keywords')) +
+                    ' to learn more about each warning.'
+                );
+                console.log(
+                    'To ignore, add ' +
+                    chalk.cyan('// eslint-disable-next-line') +
+                    ' to the line before.\n'
+                );
+            } else {
+                console.log(chalk.green('Compiled successfully.\n'));
+            }
 
-          console.log('File sizes after gzip:\n');
-          printFileSizesAfterBuild(
-              stats,
-              previousFileSizes,
-              paths.appBuild,
-              WARN_AFTER_BUNDLE_GZIP_SIZE,
-              WARN_AFTER_CHUNK_GZIP_SIZE
-          );
-          console.log();
+            console.log('File sizes after gzip:\n');
+            printFileSizesAfterBuild(
+                stats,
+                previousFileSizes,
+                paths.appBuild,
+                WARN_AFTER_BUNDLE_GZIP_SIZE,
+                WARN_AFTER_CHUNK_GZIP_SIZE
+            );
+            console.log();
 
-          const appPackage = require(paths.appPackageJson);
-          const publicUrl = paths.publicUrlOrPath;
-          const publicPath = config.output.publicPath;
-          const buildFolder = path.relative(process.cwd(), paths.appBuild);
-          printHostingInstructions(
-              appPackage,
-              publicUrl,
-              publicPath,
-              buildFolder,
-              useYarn
-          );
+            const appPackage = require(paths.appPackageJson);
+            const publicUrl = paths.publicUrlOrPath;
+            const publicPath = config.output.publicPath;
+            const buildFolder = path.relative(process.cwd(), paths.appBuild);
+            printHostingInstructions(
+                appPackage,
+                publicUrl,
+                publicPath,
+                buildFolder,
+                useYarn
+            );
         },
         err => {
-          const tscCompileOnError = process.env.TSC_COMPILE_ON_ERROR === 'true';
-          if (tscCompileOnError) {
-            console.log(
-                chalk.yellow(
-                    'Compiled with the following type errors (you may want to check these before deploying your app):\n'
-                )
-            );
-            printBuildError(err);
-          } else {
-            console.log(chalk.red('Failed to compile.\n'));
-            printBuildError(err);
-            process.exit(1);
-          }
+            const tscCompileOnError = process.env.TSC_COMPILE_ON_ERROR === 'true';
+            if (tscCompileOnError) {
+                console.log(
+                    chalk.yellow(
+                        'Compiled with the following type errors (you may want to check these before deploying your app):\n'
+                    )
+                );
+                printBuildError(err);
+            } else {
+                console.log(chalk.red('Failed to compile.\n'));
+                printBuildError(err);
+                process.exit(1);
+            }
         }
     )
     .catch(err => {
-      if (err && err.message) {
-        console.log(err.message);
-      }
-      process.exit(1);
+        if (err && err.message) {
+            console.log(err.message);
+        }
+        process.exit(1);
     });
 
 // Create the production build and print the deployment instructions.
 function build(previousFileSizes) {
-  console.log('Creating an optimized production build...');
+    console.log('Creating an optimized production build...');
 
-  const compiler = webpack(config);
-  return new Promise((resolve, reject) => {
-    compiler.run((err, stats) => {
-      let messages;
-      if (err) {
-        if (!err.message) {
-          return reject(err);
-        }
+    const compiler = webpack(config);
+    return new Promise((resolve, reject) => {
+        compiler.run((err, stats) => {
+            let messages;
+            if (err) {
+                if (!err.message) {
+                    return reject(err);
+                }
 
-        let errMessage = err.message;
+                let errMessage = err.message;
 
-        // Add additional information for postcss errors
-        if (Object.prototype.hasOwnProperty.call(err, 'postcssNode')) {
-          errMessage +=
-              '\nCompileError: Begins at CSS selector ' +
-              err['postcssNode'].selector;
-        }
+                // Add additional information for postcss errors
+                if (Object.prototype.hasOwnProperty.call(err, 'postcssNode')) {
+                    errMessage +=
+                        '\nCompileError: Begins at CSS selector ' +
+                        err['postcssNode'].selector;
+                }
 
-        messages = formatWebpackMessages({
-          errors: [errMessage],
-          warnings: [],
+                messages = formatWebpackMessages({
+                    errors: [errMessage],
+                    warnings: [],
+                });
+            } else {
+                messages = formatWebpackMessages(
+                    stats.toJson({ all: false, warnings: true, errors: true })
+                );
+            }
+            if (messages.errors.length) {
+                // Only keep the first error. Others are often indicative
+                // of the same problem, but confuse the reader with noise.
+                if (messages.errors.length > 1) {
+                    messages.errors.length = 1;
+                }
+                return reject(new Error(messages.errors.join('\n\n')));
+            }
+            if (
+                process.env.CI &&
+                (typeof process.env.CI !== 'string' ||
+                    process.env.CI.toLowerCase() !== 'false') &&
+                messages.warnings.length
+            ) {
+                // Ignore sourcemap warnings in CI builds. See #8227 for more info.
+                const filteredWarnings = messages.warnings.filter(
+                    w => !/Failed to parse source map/.test(w)
+                );
+                if (filteredWarnings.length) {
+                    console.log(
+                        chalk.yellow(
+                            '\nTreating warnings as errors because process.env.CI = true.\n' +
+                            'Most CI servers set it automatically.\n'
+                        )
+                    );
+                    return reject(new Error(filteredWarnings.join('\n\n')));
+                }
+            }
+
+            const resolveArgs = {
+                stats,
+                previousFileSizes,
+                warnings: messages.warnings,
+            };
+
+            if (writeStatsJson) {
+                return bfj
+                    .write(paths.appBuild + '/bundle-stats.json', stats.toJson())
+                    .then(() => resolve(resolveArgs))
+                    .catch(error => reject(new Error(error)));
+            }
+
+            return resolve(resolveArgs);
         });
-      } else {
-        messages = formatWebpackMessages(
-            stats.toJson({ all: false, warnings: true, errors: true })
-        );
-      }
-      if (messages.errors.length) {
-        // Only keep the first error. Others are often indicative
-        // of the same problem, but confuse the reader with noise.
-        if (messages.errors.length > 1) {
-          messages.errors.length = 1;
-        }
-        return reject(new Error(messages.errors.join('\n\n')));
-      }
-      if (
-          process.env.CI &&
-          (typeof process.env.CI !== 'string' ||
-              process.env.CI.toLowerCase() !== 'false') &&
-          messages.warnings.length
-      ) {
-        // Ignore sourcemap warnings in CI builds. See #8227 for more info.
-        const filteredWarnings = messages.warnings.filter(w => !/Failed to parse source map/.test(w));
-        if (filteredWarnings.length) {
-          console.log(
-              chalk.yellow(
-                  '\nTreating warnings as errors because process.env.CI = true.\n' +
-                  'Most CI servers set it automatically.\n'
-              )
-          );
-          return reject(new Error(filteredWarnings.join('\n\n')));
-        }
-      }
-
-      const resolveArgs = {
-        stats,
-        previousFileSizes,
-        warnings: messages.warnings,
-      };
-
-      if (writeStatsJson) {
-        return bfj
-            .write(paths.appBuild + '/bundle-stats.json', stats.toJson())
-            .then(() => resolve(resolveArgs))
-            .catch(error => reject(new Error(error)));
-      }
-
-      return resolve(resolveArgs);
     });
-  });
 }
 
 function copyPublicFolder() {
-  fs.copySync(paths.appPublic, paths.appBuild, {
-    dereference: true,
-    filter: file => file !== paths.appHtml,
-  });
+    fs.copySync(paths.appPublic, paths.appBuild, {
+        dereference: true,
+        filter: file => file !== paths.appHtml,
+    });
 }
