@@ -83,28 +83,19 @@ checkBrowsers(paths.appPath, isInteractive)
       const appName = require(paths.appPackageJson).name;
 
       const useTypeScript = fs.existsSync(paths.appTsConfig);
-      const tscCompileOnError = process.env.TSC_COMPILE_ON_ERROR === 'true';
       const urls = prepareUrls(
           protocol,
           HOST,
           port,
           paths.publicUrlOrPath.slice(0, -1)
       );
-      const devSocket = {
-        warnings: warnings =>
-            devServer.sockWrite(devServer.sockets, 'warnings', warnings),
-        errors: errors =>
-            devServer.sockWrite(devServer.sockets, 'errors', errors),
-      };
       // Create a webpack compiler that is configured with custom messages.
       const compiler = createCompiler({
         appName,
         config,
-        devSocket,
         urls,
         useYarn,
         useTypeScript,
-        tscCompileOnError,
         webpack,
       });
       // Load proxy config
@@ -115,16 +106,14 @@ checkBrowsers(paths.appPath, isInteractive)
           paths.publicUrlOrPath
       );
       // Serve webpack assets generated by the compiler over a web server.
-      const serverConfig = createDevServerConfig(
-          proxyConfig,
-          urls.lanUrlForConfig
-      );
-      const devServer = new WebpackDevServer(compiler, serverConfig);
+      const serverConfig = {
+        ...createDevServerConfig(proxyConfig, urls.lanUrlForConfig),
+        host: HOST,
+        port,
+      };
+      const devServer = new WebpackDevServer(serverConfig, compiler);
       // Launch WebpackDevServer.
-      devServer.listen(port, HOST, err => {
-        if (err) {
-          return console.log(err);
-        }
+      devServer.startCallback(() => {
         if (isInteractive) {
           clearConsole();
         }
